requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.2
django-import-export==3.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.2
numpy==1.24.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.2
pandas==2.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
